requirements.txt
streamlit
requests
pillow
transformers
torch
